Move sources-tree tests into a subfolder
